UNET-model/requirements.txt
# Install PyTorch separately with:
# pip3 install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu126
segmentation-models-pytorch>=0.3.0
opencv-python>=4.5.3
albumentations>=1.1.0
numpy>=1.20.0
matplotlib>=3.4.0
scikit-learn>=0.24.2
tqdm>=4.62.0
Pillow>=8.3.0
pycocotools>=2.0.0
seaborn>=0.11.0
pandas>=1.3.0 